Update esc.py
Since Angstrom is not RealTime OS , I was not able to read the original PWM values using the above python code. I tested using a 1.5ms input signal to the beaglebone Black GPIO pins but unfortunately, the read value's were in the range 1.2ms-1.8ms . Which is not good. 

One Solution for this was to use BBB PRU's for reading PWM values rather than using python or C++. Or to use a dedicated Microcontroller(like Atmega328) for reading PWM values and send it to BBB using I2C , SPI or UART.